refactor: OrderTable, TableGroup package 이동
