+++ src/main/java/kitchenpos/table/domain/entity/TableGroup.java
@@ -6,7 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.order.domain.value.OrderTables;
+import kitchenpos.table.domain.value.OrderTables;
 
 @Entity
 public class TableGroup {
