+++ src/main/java/kitchenpos/table/exception/NotFoundTableGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.table.exception;
 
 public class NotFoundTableGroupException extends IllegalArgumentException{
     public NotFoundTableGroupException(){
