+++ src/main/java/kitchenpos/table/service/TableGroupService.java
@@ -10,13 +10,13 @@ import kitchenpos.table.domain.entity.OrderTableRepository;
 import kitchenpos.table.domain.entity.TableGroup;
 import kitchenpos.table.domain.entity.TableGroupRepository;
 import kitchenpos.order.domain.value.OrderStatus;
-import kitchenpos.order.domain.value.OrderTables;
+import kitchenpos.table.domain.value.OrderTables;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.exception.NotFoundOrderTableException;
-import kitchenpos.order.exception.NotFoundTableGroupException;
-import kitchenpos.order.exception.OrderStatusInCookingOrMealException;
+import kitchenpos.table.exception.NotFoundTableGroupException;
+import kitchenpos.table.exception.OrderStatusInCookingOrMealException;
 import kitchenpos.table.exception.OrderTableCountException;
 import kitchenpos.table.exception.OrderTableHasTableGroupException;
 import kitchenpos.table.exception.OrderTableIsNotEmptyException;
