+++ src/test/java/kitchenpos/table/service/TableGroupServiceTest.java
@@ -16,11 +16,10 @@ import kitchenpos.table.domain.entity.OrderTableRepository;
 import kitchenpos.table.domain.entity.TableGroup;
 import kitchenpos.table.domain.entity.TableGroupRepository;
 import kitchenpos.table.domain.value.NumberOfGuests;
-import kitchenpos.order.domain.value.OrderTables;
+import kitchenpos.table.domain.value.OrderTables;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
-import kitchenpos.table.service.TableGroupService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
